storyai-backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
huggingface-hub==0.17.0
Pillow==10.0.0
requests==2.31.0
gunicorn==21.2.0
